requirements.txt
bottle==0.12.18
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.2.0
google-cloud-firestore==1.6.1
googleapis-common-protos==1.51.0
grpcio==1.26.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
